fix: altera index de relatórios para login
